underwriting-assistant/requirements.txt
# =========================
# LangChain Core (0.3.x)
# =========================
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.7
langchain-openai==0.3.5

# =========================
# Vector Database
# =========================
chromadb==0.4.24

# =========================
# Embeddings
# =========================
sentence-transformers==2.6.1

# =========================
# Data Validation
# =========================
pydantic==2.7.1

# =========================
# UI & Utilities
# =========================
streamlit==1.38.0
python-dotenv==1.0.1
